Turned the jagged array into a flat one.
Changed 2 lists into a pair of arrays and a single length counter. Got rid of Add, RemoveRange, and Clear calls by doing so.
I considered stackalloc Span<T>, but the arrays can get quite large and not fit the stack.

In GetSegmentCost I pulled `tau2 - tau1` out of the loop and also made indexing simpler.
In GetPartialSums I removed some array accesses by using a variable to store the previous value, and saved some writes by calculating the value before assigning it.

Added curly braces

Is there a reason it's not a static class?