@AndreyAkinshin 2019-10-24
@jeanbern thanks for the PR!
Do you have any benchmarks that show noticeable performance impact from your changes?