@jeanbern 2019-10-24
Running unit tests as benchmarks:

Before:

| Method |           Mean |         Error |        StdDev |
|------- |---------------:|--------------:|--------------:|
|  Test1 |       8.350 us |     0.1015 us |     0.0848 us |
|  Test2 |   3,554.749 us |    11.8809 us |    10.5321 us |
|  Test3 |       4.592 us |     0.0285 us |     0.0253 us |
|  Test4 |      16.173 us |     0.0864 us |     0.0808 us |
|  Test5 |   7,649.757 us |    39.3496 us |    32.8587 us |
|  Test6 | 492,407.260 us | 3,830.0268 us | 3,582.6095 us |


After:

| Method |           Mean |         Error |        StdDev |
|------- |---------------:|--------------:|--------------:|
|  Test1 |       6.354 us |     0.0446 us |     0.0372 us |
|  Test2 |   2,960.071 us |    22.6944 us |    20.1180 us |
|  Test3 |       3.463 us |     0.0089 us |     0.0074 us |
|  Test4 |      12.128 us |     0.0722 us |     0.0640 us |
|  Test5 |   4,802.458 us |    33.9739 us |    31.7792 us |
|  Test6 | 301,802.931 us | 1,945.7635 us | 1,624.8007 us |