@AndreyAkinshin 2019-11-26
@jeanbern sorry for the late review, I was extremely busy the last month =(

I pushed some fixed into your branch, here is an overview:
* Remove curly brackets from single-line statement to be consistent with the common code style in BenchmarkDotNet
* Improve readability: add some comments here and there, rename some variables, etc.
* Rollback the `do/while` loop in `GetPartialSums` back to a simple for. According to my measurements, the `do/while` doesn't provide a noticeable performance improvement, but it significantly reduces the code readability

Let me know what do you think. If you are OK with the changes, I'm ready to merge the branch.

In general, the PR is great! Replacing two-dimensional array by a single-dimensional one produced a very good overall performance impact!

> Is there a reason it's not a static class?

Yes, I'm going to implement a few additional algorithms for changepoint detection and define a common interface for them.