@jeanbern 2019-11-26
@AndreyAkinshin Looks good to me, thank you for taking the time to go over it and explain your changes. I'll keep the comments in mind when I submit PR in the future.